немного отредактировал структуру приложения, создал компонент Api.js
