--- src/components/App/ThemeContext.js
@@ -1,4 +1,0 @@
-// import { createContext } from 'react'
-
-// export const DEFAULT_THEME = 'light'
-// export const ThemeContext = createContext(DEFAULT_THEME)
