+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from "redux";
-// import { postReducer } from "./post";
 import { themeReducer } from "./theme";
-import { projectsReducer } from "./projects";
-import { tasksReducer } from "./tasks";
+import { projectsReducer, tasksReducer } from "./projects_tasks";
 
 export const rootReducer = combineReducers({
   theme: themeReducer,
