+++ src/reducers/projects_tasks.js
@@ -1,8 +1,8 @@
-import { PROJECT_ADD, PROJECT_TASK_ADD } from '../actions/projects/projects'
-import { projects } from '../components/NormalizeProjects/NormalizeProjects'
-import { NormalizeProjects } from '../components/NormalizeProjects/NormalizeProjects'
+import { PROJECT_ADD, PROJECT_TASK_ADD, CHANGE_STATUS, TASK_ADD } from '../actions/projects_tasks'
+import { projects } from '../components/NormalizeState'
+import NormalizeState from '../components/NormalizeState'
 
-const { projectsById, tasksById } = NormalizeProjects(projects)
+const { projectsById, tasksById } = NormalizeState(projects)
 
 const initialState = {
   projects: projectsById,
@@ -39,6 +39,38 @@ export const projectsReducer = (state = initialState, action) => {
       return state;
   }
 }
+
+export const tasksReducer = (state = initialState, action) => {
+    switch (action.type) {
+        case TASK_ADD: {
+            const taskId = Object.keys(state.tasks).length + 1
+            const newTasks = { ...state.tasks }
+            const newTask = {
+                id: taskId,
+                name: action.name,
+                description: action.description,
+                completed: false
+            }
+            newTasks[taskId] = newTask
+            return {
+                ...state,
+                tasks: newTasks
+            }
+        }
+        case CHANGE_STATUS: {
+            const taskId = action.id
+            const newTasks = { ...state.tasks }
+            newTasks[taskId].completed = !action.completed
+            return {
+                ...state,
+                tasks: newTasks
+            }
+        }
+        default:
+            return state;
+    }
+}
+
 
 
 // export const projects = {
